backend/requirements.txt
bson
wsgiserver
aniso8601==3.0.2
atomicwrites==1.2.1
attrs==18.2.0
click==6.7
colorama==0.3.9
Flask==1.0.2
Flask-JWT==0.3.2
Flask-RESTful==0.3.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.3.0
pluggy==0.7.1
py==1.6.0
PyJWT==1.4.2
pytest==3.8.0
pytz==2018.5
six==1.11.0
Werkzeug==0.14.1
